Rename chalicelib to src
Also, chalicelib is supposed to be the root of a package hierarchy, not a package itself.
