+++ webservice/app.py
@@ -1,9 +1,9 @@
 import ast
-from chalicelib import config
+import config
 import logging.config
 import os
-from chalicelib.responseobjects.elastic_request_builder import BadArgumentException, ElasticTransformDump as EsTd
-from chalicelib.responseobjects.utilities import json_pp
+from responseobjects.elastic_request_builder import BadArgumentException, ElasticTransformDump as EsTd
+from responseobjects.utilities import json_pp
 from chalice import Chalice
 
 ENTRIES_PER_PAGE = 10
