+++ webservice/src/config/autocomplete_mapping_config.json
