+++ webservice/src/config/mapping_config.json
