+++ webservice/src/responseobjects/hca_response_v5.py
@@ -1,6 +1,6 @@
 #!/usr/bin/python
 import abc
-from chalicelib.responseobjects.utilities import json_pp
+from responseobjects.utilities import json_pp
 from chalice import Response
 from collections import defaultdict
 from collections import OrderedDict
